made a real function finally... im not totally useless

+++ src/main.py
@@ -2,19 +2,33 @@ from textnode import TextNode
 import os
 import shutil
 
-def copy_from_public():
-    files = os.listdir("static")
-    path = "static"
+def copy_from_static(path, dest):
+    files = os.listdir(path)
+    root = 'static'
+    src = path
 
-    shutil.rmtree("public")
-    os.mkdir("public")
+    if path is root:
+        shutil.rmtree("public")
+        os.mkdir("public")
 
     for file in files:
-        path += f'/{file}'
-        print(os.path.isfile(path), '\n')
-        path = 'static'
+        src += f'/{file}'
+        dest_path = dest
+        if os.path.isfile(src):
+            print(src, 'src')
+            print(dest_path, 'dest\n\n')
+            shutil.copy(src, dest_path)
+        else:
+            print(src, 'src not file')
+            dest_path += f'/{file}'
+            os.mkdir(dest_path)
+            print(dest_path, 'dest not file\n\n')
+            copy_from_static(src, dest_path)
+
+
+        src = 'static'
 
 def main():
-    copy_from_public()
+    copy_from_static("static", "public")
 
 main()
